sap OCB설정
git-svn-id: https://skvalley.com/svn/repos/storeplatform/trunk/sac/storeplatform-sac@74844 01f5c84a-aead-4b29-b0f1-7c54dbeec05f

+++ src/main/java/com/skplanet/storeplatform/sac/purchase/order/service/PurchaseOrderPolicyServiceImpl.java
@@ -1207,7 +1207,7 @@ public class PurchaseOrderPolicyServiceImpl implements PurchaseOrderPolicyServic
 			int resultMaxPer = (int) (Double.parseDouble(resultPhonePolicyInfo[2]));
 
 			double phoneAvailAmt = resultMaxAmt;
-			double phoneAvailPer = resultMaxPer;
+			int phoneAvailPer = resultMaxPer;
 
 			if (phonePos >= 0) {
 				int sepPos = paymentAdjustInfo.indexOf(";", phonePos);
